# docs/api/arx_backend_api_spec.yaml
openapi: 3.0.3
info:
  title: Arx Backend API
  description: |
    Comprehensive API for the Arx Backend service providing BIM object management,
    asset inventory, CMMS integration, maintenance workflows, and compliance reporting.

    ## Authentication
    The API uses JWT Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Rate Limiting
    - Public endpoints: 100 requests/second, burst of 200
    - Authenticated endpoints: 50 requests/second, burst of 100

    ## Error Handling
    All endpoints return standard HTTP status codes with detailed error messages in JSON format.
  version: 1.0.0
  contact:
    name: Arx Development Team
    email: dev@arxos.com
  license:
    name: Proprietary
    url: https://arxos.com/license
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.arxos.com/api
    description: Production server
...
paths:
...
  /buildings:
    get:
      tags:
        - Buildings
      summary: List buildings
      description: Retrieve a list of all buildings
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, maintenance]
      responses:
        '200':
          description: List of buildings
          content:
            application/json:
              schema:
                type: object
                properties:
                  buildings:
                    type: array
                    items:
                      $ref: '#/components/schemas/Building'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Buildings
      summary: Create building
      description: Create a new building
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                address:
                  type: string
                city:
                  type: string
                state:
                  type: string
                zip_code:
                  type: string
                country:
                  type: string
                  default: "USA"
                building_type:
                  type: string
                total_floors:
                  type: integer
                  minimum: 1
                total_area:
                  type: number
                  minimum: 0
                year_built:
                  type: integer
      responses:
        '201':
          description: Building created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Building'
  /buildings/{id}:
    get:
      tags:
        - Buildings
      summary: Get building
      description: Retrieve a specific building by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Building details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Building'

    put:
      tags:
        - Buildings
      summary: Update building
      description: Update an existing building
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Building'
      responses:
        '200':
          description: Building updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Building'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Building:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, inactive, maintenance, archived]
        owner_id:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean